0722 : 오큰수 (stack)

+++ src/week19/BOJ28137_뭐라고안들려.java
@@ -6,9 +6,9 @@ import java.io.*;
 // 점을 넣은 함수의 y절편이 같다면 같은 선 위에 있음
 public class BOJ28137_뭐라고안들려 {
 	static int N;
-	static double INCL;
+	static long INCL;
 	static long cnt;
-	static Map<Double, Double> map = new HashMap<>();
+	static Map<Long, Long> map = new HashMap<>();
 
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -16,17 +16,17 @@ public class BOJ28137_뭐라고안들려 {
 
 		st = new StringTokenizer(br.readLine());
 		N = Integer.parseInt(st.nextToken());
-		INCL = Double.parseDouble(st.nextToken());
+		INCL = Long.parseLong(st.nextToken());
 
 		for (int i = 0; i < N; i++) {
 			st = new StringTokenizer(br.readLine());
-			double x = Double.parseDouble(st.nextToken());
-			double y = Double.parseDouble(st.nextToken());
+			long x = Long.parseLong(st.nextToken());
+			long y = Long.parseLong(st.nextToken());
 
 			addToMap((-x) * INCL + y);
 		}
 
-		for (double key : map.keySet()) {
+		for (long key : map.keySet()) {
 			perm(map.get(key));
 		}
 
@@ -34,12 +34,12 @@ public class BOJ28137_뭐라고안들려 {
 
 	}
 
-	private static void perm(double val) {
+	private static void perm(long val) {
 		cnt += (val) * (val - 1);
 	}
 
-	private static void addToMap(double yIntercept) {
-		Double getCnt = map.getOrDefault(yIntercept, (double) 0);
+	private static void addToMap(long yIntercept) {
+		Long getCnt = map.getOrDefault(yIntercept, (long) 0);
 		map.put(yIntercept, getCnt + 1);
 	}
